refactor(middleware): rename middleware files and update imports
- rename auth.js → auth.middleware.js
- rename error.js → error.middleware.js
- rename rateLimit.js → rateLimit.middleware.js
- update all related imports in routes and app.js
