+++ src/routes/sessions.routes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const auth = require('../middleware/auth');
+const auth = require('../middleware/auth.middleware');
 const {
   createSession,
   latestSession,
